Add Receive Run Queue
RQ in each kthread is a single procuder (the kthread itself) and
multi-consumer queue.
RRQ is the opposite that every kthread needs to lock to enqueue a task
and update the head, but only the owner kthread dequeues the received
task targetted to run with this kthread and updates the tail without
locking.