Correct number of running threads counting
In an earlier commit, `poll_thread()` was removed to avoid dangling
process reference, but then the number of running threads (nrts) is not
counted properly, leading to park signal from the runtime is never
received and processed.
This commit get `poll_thread()` and `unpoll_thread()` back to the right
place and updates `nrts` correctly.