Set affinity to kthread instead of core
When the application does not know what are the cores available, it is
really difficult to make use of the thread_throw() interface to set
affinity.
This commit switches to set affinity to kthreads, which have gotten the
hwalloc daemon's best effort to stay on the same core already. This way
the application can create tasks with affinity indirectly set to core,
but it is more robust and convenient to do so.
The new issue is that with kthread, we cannot assume two kthreads are
closer to each other even though they are on two core sharing the same
physical cores or local cache cluster.